backend/requirements-dev.txt
# Development and optional dependencies
-r requirements.txt

# Background task processing
celery==5.3.4
redis==5.0.1

# Logging and monitoring
structlog==23.2.0
psutil==5.9.6

# API development
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Testing dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-benchmark==4.0.0

# Test data generation
factory-boy==3.3.0
Faker==20.1.0
freezegun==1.2.2
responses==0.24.1

# Code quality and formatting
black==23.11.0
ruff==0.1.7
mypy==1.7.1
pre-commit==3.5.0

# PDF testing utilities
reportlab==4.0.7
PyPDF2==3.0.1
pillow==10.1.0
numpy==1.25.2
opencv-python-headless==4.8.1.78
pandas==2.1.3
camelot-py==0.11.0
tabula-py==2.9.0
pytesseract==0.3.10
langdetect==1.0.9
pytest-env==1.1.3